Begin work on focus
Focus is now switched between different modules when they are
moused-over. Modules can also retain focus through use of a boolean
value. In the demo program, there are two editor modules that are
stacked on top of each other. Whichever one the mouse is in is in focus.
If panning the module by pressing the scroll wheel, the editor will
retain focus until the mouse is released.

Other changes: panning was given its own functions that are called so
that it is easier to modify them.