Get rid of drawSelf(), replace with draw()
draw() used to make the parent re-draw all child modules. I couldn't
think of when this would ever be necessary, so I deleted drawSelf() and
made draw() do what drawSelf() used to do. Now draw() will cause the
module to re-draw and then tell its parent to update with the new render
of the child module.