Edit ModulePane's resize method
ModulePane now resizes right. Although the normal
Container.resizeComponent() method leads to a loss of info in the child
modules, the ModulePane needs to have all modules separated by 3 by
default. However, right now the method will not keep user-resized sizes
when the window is resized. This needs to be changed so that the ratios
all work out right.