Update various ModulePane rendering code
The two ModulePane classes have updated rendering code. First, resizing
is slightly more user-friendly where moving the arrow outside the
possible resize doesn't cause the resizing to be later offset. However,
this has caused several annoying bugs where the mouse is moved quickly
when at the edge of a resize. ModulePanes have also been updated to
retain their focus when being resized so that their parent ModulePanes
won't do anything while their children are resizing. Also, the
mouseMove() method of a child is called when it goes out of focus so it
can update (for example, moving the resize arrow out of the visible
window)