Update DynModule rendering
DynModule now actually draws when its child wants it to draw. It also
draws a nice looking border around it which will be used for the real
purpose of the DynModule. The image in the lower right corner shows that
it can be split. TODO: Come up with good UI for splitting at top / left.
I think the best way to do this may be with simply another image.