Update DynModule; misc changes
First, each module can now have its own specialized minimum size which
both HorizontalModulePane and VerticalModulePane will read. I also fixed
passing of mouseData to actually work correctly - before it appeared
that it worked correctly in all my tests because of strange
circumstances. DynModule now displays a little arrow when it is possible
to split it at the bottom-right corner. All the borders around the sides
except at the bottom were abolished. It still needs to be able to be
split at the bottom (and maybe eventually at the side) and be
dragged-split in the upper left hand corner. Also, methods need to be
implemented to actually split the container.