Set up further architecture for program
More architecture has been set up for the basic system of vague. The
Vector class simply stores two integers, x and y, which are presumably
related to a rectangular coordinate system. The ImageLoader class loads
images and also stores the object IMG which is a container for all image
resources. The Module class has been very well set up for the Module
system, but it still needs to be able to render the parent.

The Program class with the entry point of the application initializes
the image resources container. In the future, it will also handle
loading for any other resources. I would also like to have a loading
screen.

The Window class also now sets its default close operation to
EXIT_ON_CLOSE. This is nice so that the application actually exits when
it is closed.