Fix most problems with the module system
Everything is initialized in the right order in Window, and all of the
Module functions seem to work properly. Now I can re-implement the
container classes. The Doodle class works fine if you want to draw weird
pictures that you can't save.

So yay this program is actually a drawing program right now!