Fix module problem; Pass container events
The basic Container class now both inherits from Module and passes on
all events to its subclasses. A problem with doRenderComps() was fixed
where if the Module had a size with either parameter of zero, then the
normal image was done correctly but the buffer to copy old data was not.