Implement container resizing
Resizing of containers now works as well as it is going to. I need to
implement a method that will calculate the height of the Container if a
width and height are not provided - this simply requires generating a
width and height, and then checking to see if any components have a
right / bottom that extends outside the current width / height, and
simply extending the Container if that is the case.  The redraw() method
in the Exchange class is called once the program is run such that an
initial graphical state is guaranteed.

The Percents class needed to cast integers to doubles when calculating
percentages so that has been fixed.

The TestModule class is simply that, a test Module that I am using to
test various features. Next I am going to make sure focus change works
right.

I think I need to make an onFocus() and maybe onUnfocus() method for the
Module class so that that works out nicely.