Fixed container initialization without specified size
Before the container would not initialize properly without a specified
size. In order to simply initialize a Module to a certain size, the
initialize() method was added but should ONLY be called in the
constructor. The Module also now checks to make sure it is being resized
to a valid size, and will not resize smaller than 0,0. (Note: this may
eventually cause weird problems where the app is breaking but silently;
hopefully this won't happen.)