Update cosmetics
The Window now starts, in general, in the center of the screen. A popup
was added that displays when the application is loading resources so
that it is clear. The resources system was slightly reworked so that
ImageLoader is now called Resources and that the loadResources() method
must be called.