Create WorkTool and make it do things
The WorkTool class defines a sort of window that the user can create
(and in the future, will be able to edit) inside the Workspace. Right
now the user can create very basic WorkTools. Multiple modifications
were made to other classes in order to accommodate the Workspace system,
such as Container's members being made public.