Make basic WorkTool removal
WorkTools can now be removed by clicking on them - in the future this
will be caused by clicking on a specialized dismiss button (such as an x
or something.) This also required making the removeChild(Module) method
in Container, which will remove the specified child Module from the
Container. WorkTool passes 'this'. This, of course, requires casting to
Container() of the parent. It is safe using instanceof.