Change Key control system
Now all key controls need to be updated to use the Controls.bank to
determine whether keys are pressed. This will allow for a centralized
key status system and also changeable key actions in the future. The
onKeyDown() method should probably also be changed to not pass a
KeyEvent because it is not needed, but that has yet to be seen.