Fix / update WorkTool moving/resizing
A few graphical bugs were updated, such as the state of the controls not
being updated and an error where the buffer in Workspace would contain a
version of WorkTools that had previously stopped resizing that had a
white/red outline rather than black. WorkTools are now outlined in red
when they are being moved / resized and they are invalid. Also, an error
where the resize reset would not work correctly was fixed.