Make refresh button; fix focus issue
The refresh button in the upper-right corner of a WorkTool causes it to
reset to an initial state where a new control can be selected. I also
fixed a problem where buttons could stay focused inside a SmartMenu if
the mouse was moved out of the menu quickly; now if the mouse is moved
out of the menu, all the buttons are make unpressed if need be.