Create zooming and view reset
Zooming is done by scaling the canvas by powers of two. The maximum
scale is 2^7, or 128, and the minimum scale is 2^-3, or 1/8. This is
mostly due to memory limits and things, the minimum scale is based on
logic or something. I may make it so that the max zoom is based on the
size of the image.

I fixed the Window class to actually listen for mouse wheel events,
which happens to be important while zooming.

Graphical changes to the canvas will have to be managed carefully to
keep the program running fast enough - maybe the Editor can just draw
changes directly to its buffer, already scaled. However, then layers and
transparency and things begin to become problems.