Fix slight graphical bug; add toggleable grid lines
There was a slight graphical bug when using 8 to reset position / zoom -
one of the controls would cause the background buffer to not be
re-created (I don't remember which control....)

I also added grid lines that can be toggled to show the grid of pixels -
these will only appear for scale >= 4x because otherwise they don't look
very good.

Grid lines are toggled with shift-S.

I might changed the control structure to use "Chords" so that in the
future all controls will be completely customizable.