Further work on system
Vector class moved to module package; accounts for many changes

The new Module system is growing but should be made pretty modern very
soon. The main problem right now is that each Module, as of now, only
has access to a single paint() method. This does not allow for efficient
drawing when only a small portion of a Module needs to be redrawn.

That is why I will soon implement a more localized drawing system where
Modules can request very specific areas of access to the canvas.

A big change in the new Module system is that Modules no longer need to
re-draw their parent when they have updated their graphical state. All
changes are directly reflected in the top-level window. This also makes
all the buffers used completely obsolete.