Begin to finalize new painting
A few things were tweaked, including the fact that there before was
actually no call to a paint method when needed. Hopefully now all the
necessary framework is there for painting the entire Module using the
Module.repaint() method.

I also began implementing a callback such that painting can be done on a
partial basis. This is not quite finished as Modules do not have any way
to access this as of now.

Things to fix:
-Modules need to always have a handle to the Window.
-Modules need a way to have access to partial painting.