Update Modules to hopefully work with new painting
I have built Modules to require a handle to a Window to be passed in the
constructor. The handle is final. (note to self - as the handle is
final, it doesn't have to be private anymore - right?) Modules now also
have the ability to request partial redrawing with a callback to when
the graphical handle is returned, so hopefully it should be possible to
start fixing the currently built Modules in the near future.