Start migrating code
I have migrated much of the code to the new system. However, I foresee a
very real problem: various code execution issues. Essentially, if too
many things are trying to repaint at once, most of them won't actually
be re-painted.

I am going to have to rework the system further.

As it stands, the current code does essentially nothing but open a blank
window. There is a lot of stuff to be done before this works again.