Slightly improve new Context
Now each ContextView keeps a local copy of all the values that it needs
to calculate new positions, so this should make this code slightly
easier to read and should make redrawing the screen while drawing
slightly easier to implement.

I basically need to implement a projection system, allowing me to
project coordinates to the context and back.