Make pencil tool retain focus; change min/max size
The pencil tool retains focus while drawing. I still cannot understand
why it sometimes glitches out and get really big. I think I might have
an idea of how to debug it, but it will be hard because I'm not really
sure what the procedure is to cause the glitch.